Refactor meteor-vite directory structure
