+++ npm-packages/meteor-vite/package.json
@@ -7,7 +7,7 @@
     "./client": "./dist/client/index.mjs"
   },
   "scripts": {
-    "start": "ts-node-dev src/dev/debug-runner.ts",
+    "start": "ts-node-dev src/bin/dev/debug-runner.ts",
     "watch": "tsup --watch",
     "build": "tsup",
     "test": "echo \"Error: no test specified\" && exit 1"
