--- npm-packages/meteor-vite/src/client/index.ts
@@ -1,1 +1,0 @@
-export { validateStub } from './ClientStubValidation';
