--- npm-packages/meteor-vite/src/dev/ViteServer.ts
@@ -1,16 +1,0 @@
-import FS from 'fs';
-import Path from 'path';
-import { createServer } from 'vite';
-import { MeteorViteStubs } from '../vite/MeteorViteStubs';
-
-const projectRoot = Path.join(__dirname, '../../.../../../../examples/vue/');
-
-export default createServer({
-    root: projectRoot,
-    plugins: [
-        MeteorViteStubs({
-            meteorPackagePath: Path.join(projectRoot, '.meteor', 'local', 'build', 'programs', 'web.browser', 'packages'),
-            projectJsonContent: JSON.parse(FS.readFileSync(Path.join(projectRoot, 'package.json'), 'utf-8')),
-        })
-    ]
-});
