--- npm-packages/meteor-vite/src/index.ts
@@ -1,1 +1,0 @@
-export { MeteorViteStubs } from './vite/MeteorViteStubs';
