+++ npm-packages/meteor-vite/src/meteor/client/index.ts
@@ -1,0 +1,1 @@
+export { validateStub } from './ValidateStub';
