+++ npm-packages/meteor-vite/src/vite/plugin/index.ts
@@ -1,0 +1,1 @@
+export { MeteorStubs } from './MeteorStubs';
